Move back, rip out type hints
