+++ src/attr/__init__.py
@@ -10,10 +10,11 @@ from typing import Callable, Literal, Protocol
 from . import converters, exceptions, filters, setters, validators
 from ._cmp import cmp_using
 from ._config import get_run_validators, set_run_validators
-from ._funcs import asdict, assoc, astuple, has, resolve_types
+from ._funcs import asdict, assoc, astuple, has, inspect, resolve_types
 from ._make import (
     NOTHING,
     Attribute,
+    ClassProps,
     Converter,
     Factory,
     _Nothing,
@@ -26,7 +27,6 @@ from ._make import (
     validate,
 )
 from ._next_gen import define, field, frozen, mutable
-from ._props import ClassProps, inspect
 from ._version_info import VersionInfo
