+++ src/attr/__init__.pyi
@@ -20,8 +20,6 @@ from . import filters as filters
 from . import setters as setters
 from . import validators as validators
 from ._cmp import cmp_using as cmp_using
-from ._funcs import inspect as inspect
-from ._props import ClassProps as ClassProps
 from ._typing_compat import AttrsInstance_
 from ._version_info import VersionInfo
 from attrs import (
